usuarios/pyproject.toml
[tool.poetry]
name = "users"
version = "0.1.0"
description = "backend for users"
authors = ["luciakasman <lucia.kasman@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "^0.103.2"
uvicorn = "^0.23.2"
pydantic = {extras = ["email"], version = "^2.4.2"}
motor = "^3.3.1"
email-validator = "^2.1.0.post1"
sqlmodel = "^0.0.14"
sqlalchemy = "^2.0.25"
psycopg2-binary = "^2.9.9"
asyncpg = "^0.29.0"
PyJWT = "^2.8.0"
python-decouple = "^3.8"
python-multipart = "^0.0.6"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
python-dotenv = "^1.0.1"

[tool.poetry.group.dev.dependencies]
black = "^23.9.1"
pylint = "^3.0.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
